Rename metrics context to datapoints
