+++ pkg/oteltransformationlanguage/contexts/ottldatapoints/datapoints_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package ottlmetrics
+package ottldatapoints
 
 import (
 	"testing"
