+++ processor/transformprocessor/internal/metrics/func_convert_sum_to_gauge.go
@@ -17,13 +17,13 @@ package metrics // import "github.com/open-telemetry/opentelemetry-collector-con
 import (
 	"go.opentelemetry.io/collector/pdata/pmetric"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/oteltransformationlanguage/contexts/ottlmetrics"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/oteltransformationlanguage/contexts/ottldatapoints"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/oteltransformationlanguage/ottl"
 )
 
 func convertSumToGauge() (ottl.ExprFunc, error) {
 	return func(ctx ottl.TransformContext) interface{} {
-		mtc, ok := ctx.(ottlmetrics.TransformContext)
+		mtc, ok := ctx.(ottldatapoints.TransformContext)
 		if !ok {
 			return nil
 		}
