+++ processor/transformprocessor/internal/metrics/processor.go
@@ -22,7 +22,7 @@ import (
 	"go.opentelemetry.io/collector/pdata/pmetric"
 	"go.uber.org/zap"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/oteltransformationlanguage/contexts/ottlmetrics"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/oteltransformationlanguage/contexts/ottldatapoints"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/oteltransformationlanguage/ottl"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/transformprocessor/internal/common"
 )
@@ -35,8 +35,8 @@ type Processor struct {
 func NewProcessor(statements []string, functions map[string]interface{}, settings component.ProcessorCreateSettings) (*Processor, error) {
 	ottlp := ottl.NewParser(
 		functions,
-		ottlmetrics.ParsePath,
-		ottlmetrics.ParseEnum,
+		ottldatapoints.ParsePath,
+		ottldatapoints.ParseEnum,
 		common.NewOTTLLogger(settings.Logger),
 	)
 	queries, err := ottlp.ParseQueries(statements)
@@ -77,28 +77,28 @@ func (p *Processor) ProcessMetrics(_ context.Context, td pmetric.Metrics) (pmetr
 
 func (p *Processor) handleNumberDataPoints(dps pmetric.NumberDataPointSlice, metric pmetric.Metric, metrics pmetric.MetricSlice, is pcommon.InstrumentationScope, resource pcommon.Resource) {
 	for i := 0; i < dps.Len(); i++ {
-		ctx := ottlmetrics.NewTransformContext(dps.At(i), metric, metrics, is, resource)
+		ctx := ottldatapoints.NewTransformContext(dps.At(i), metric, metrics, is, resource)
 		p.callFunctions(ctx)
 	}
 }
 
 func (p *Processor) handleHistogramDataPoints(dps pmetric.HistogramDataPointSlice, metric pmetric.Metric, metrics pmetric.MetricSlice, is pcommon.InstrumentationScope, resource pcommon.Resource) {
 	for i := 0; i < dps.Len(); i++ {
-		ctx := ottlmetrics.NewTransformContext(dps.At(i), metric, metrics, is, resource)
+		ctx := ottldatapoints.NewTransformContext(dps.At(i), metric, metrics, is, resource)
 		p.callFunctions(ctx)
 	}
 }
 
 func (p *Processor) handleExponetialHistogramDataPoints(dps pmetric.ExponentialHistogramDataPointSlice, metric pmetric.Metric, metrics pmetric.MetricSlice, is pcommon.InstrumentationScope, resource pcommon.Resource) {
 	for i := 0; i < dps.Len(); i++ {
-		ctx := ottlmetrics.NewTransformContext(dps.At(i), metric, metrics, is, resource)
+		ctx := ottldatapoints.NewTransformContext(dps.At(i), metric, metrics, is, resource)
 		p.callFunctions(ctx)
 	}
 }
 
 func (p *Processor) handleSummaryDataPoints(dps pmetric.SummaryDataPointSlice, metric pmetric.Metric, metrics pmetric.MetricSlice, is pcommon.InstrumentationScope, resource pcommon.Resource) {
 	for i := 0; i < dps.Len(); i++ {
-		ctx := ottlmetrics.NewTransformContext(dps.At(i), metric, metrics, is, resource)
+		ctx := ottldatapoints.NewTransformContext(dps.At(i), metric, metrics, is, resource)
 		p.callFunctions(ctx)
 	}
 }
